refactor: update import paths for DashboardContext and clean up related components
